docs: ADR on value providers type discovery
prompt: The thing that bothers me is that ValueProviders are meant to be implemented as generics, while the real value type cannot be known at the time of graph construction, at least for some Providers. How can this problem be solved? I don't like the idea of specifying concrete types in decision table's file, because that would couple logic to the implementation. Help me find the solution.